backend/src/requirements/_base.txt
Django==3.2.23
paho-mqtt==1.6.1
djangorestframework==3.14.0
django-cors-headers==3.14.0
PyJWT==2.8.0
Pillow==10.1.0
openpyxl==3.1.2
gunicorn==21.2.0
django-import-export==3.3.6
